move utils from app
