+++ packages/frontend2/src/app/asset-risks/[address]/_components/table/risk-details.tsx
@@ -1,8 +1,8 @@
 import Image from 'next/image'
 import Link from 'next/link'
-import { formatAddress } from '~/app/utils/format-address'
 import OutLinkIcon from '~/icons/outlink.svg'
 import { cn } from '~/utils/cn'
+import { formatAddress } from '~/utils/format-address'
 import CriticalIcon from '../../_assets/critical-badge.svg'
 import WarningIcon from '../../_assets/warning-badge.svg'
 import { type Risk } from '../../page'
