+++ packages/frontend2/src/utils/format-number.ts
